refactor(acp): retain extracted command ownership during refresh
